Neuron_Project/requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0
tifffile>=2020.9.3
PyOpenGL>=3.1.5
PyOpenGL-accelerate>=3.1.5
PySide6>=6.6.0
scikit-image>=0.19.0
scipy>=1.7.0 